Move Compose API in api/compose.py module; DRY; fixes
