+++ counterparty-core/counterpartycore/lib/api/routes.py
@@ -1,5 +1,4 @@
-from counterpartycore.lib import transaction
-from counterpartycore.lib.api import queries, util
+from counterpartycore.lib.api import compose, queries, util
 from counterpartycore.lib.backend import addrindexrs, bitcoind
 
 
@@ -31,9 +30,9 @@ ROUTES = util.prepare_routes(
         "/v2/blocks/<int:block_index>/sweeps": queries.get_sweeps_by_block,
         ### /transactions ###
         "/v2/transactions": queries.get_transactions,
-        "/v2/transactions/info": transaction.info,
-        "/v2/transactions/<tx_hash>/info": transaction.info_by_tx_hash,
-        "/v2/transactions/unpack": transaction.unpack,
+        "/v2/transactions/info": compose.info,
+        "/v2/transactions/<tx_hash>/info": compose.info_by_tx_hash,
+        "/v2/transactions/unpack": compose.unpack,
         "/v2/transactions/<int:tx_index>": queries.get_transaction_by_tx_index,
         "/v2/transactions/<tx_hash>": queries.get_transaction_by_hash,
         "/v2/transactions/<int:tx_index>/events": queries.get_events_by_transaction_index,
@@ -74,25 +73,25 @@ ROUTES = util.prepare_routes(
         "/v2/addresses/<address>/fairmints": queries.get_fairmints_by_address,
         "/v2/addresses/<address>/fairmints/<asset>": queries.get_fairmints_by_address_and_asset,
         ### /addresses/<address>/compose/ ###
-        "/v2/addresses/<address>/compose/bet": transaction.compose_bet,
-        "/v2/addresses/<address>/compose/broadcast": transaction.compose_broadcast,
-        "/v2/addresses/<address>/compose/btcpay": transaction.compose_btcpay,
-        "/v2/addresses/<address>/compose/burn": transaction.compose_burn,
-        "/v2/addresses/<address>/compose/cancel": transaction.compose_cancel,
-        "/v2/addresses/<address>/compose/destroy": transaction.compose_destroy,
-        "/v2/addresses/<address>/compose/dispenser": transaction.compose_dispenser,
-        "/v2/addresses/<address>/compose/dividend": transaction.compose_dividend,
-        "/v2/addresses/<address>/compose/issuance": transaction.compose_issuance,
-        "/v2/addresses/<address>/compose/mpma": transaction.compose_mpma,
-        "/v2/addresses/<address>/compose/order": transaction.compose_order,
-        "/v2/addresses/<address>/compose/send": transaction.compose_send,
-        "/v2/addresses/<address>/compose/sweep": transaction.compose_sweep,
-        "/v2/addresses/<address>/compose/dispense": transaction.compose_dispense,
-        "/v2/addresses/<address>/compose/fairminter": transaction.compose_fairminter,
-        "/v2/addresses/<address>/compose/fairmint": transaction.compose_fairmint,
-        "/v2/addresses/<address>/compose/attach": transaction.compose_attach,
-        "/v2/utxos/<utxo>/compose/detach": transaction.compose_detach,
-        "/v2/utxos/<utxo>/compose/movetoutxo": transaction.compose_movetoutxo,
+        "/v2/addresses/<address>/compose/bet": compose.compose_bet,
+        "/v2/addresses/<address>/compose/broadcast": compose.compose_broadcast,
+        "/v2/addresses/<address>/compose/btcpay": compose.compose_btcpay,
+        "/v2/addresses/<address>/compose/burn": compose.compose_burn,
+        "/v2/addresses/<address>/compose/cancel": compose.compose_cancel,
+        "/v2/addresses/<address>/compose/destroy": compose.compose_destroy,
+        "/v2/addresses/<address>/compose/dispenser": compose.compose_dispenser,
+        "/v2/addresses/<address>/compose/dividend": compose.compose_dividend,
+        "/v2/addresses/<address>/compose/issuance": compose.compose_issuance,
+        "/v2/addresses/<address>/compose/mpma": compose.compose_mpma,
+        "/v2/addresses/<address>/compose/order": compose.compose_order,
+        "/v2/addresses/<address>/compose/send": compose.compose_send,
+        "/v2/addresses/<address>/compose/sweep": compose.compose_sweep,
+        "/v2/addresses/<address>/compose/dispense": compose.compose_dispense,
+        "/v2/addresses/<address>/compose/fairminter": compose.compose_fairminter,
+        "/v2/addresses/<address>/compose/fairmint": compose.compose_fairmint,
+        "/v2/addresses/<address>/compose/attach": compose.compose_attach,
+        "/v2/utxos/<utxo>/compose/detach": compose.compose_detach,
+        "/v2/utxos/<utxo>/compose/movetoutxo": compose.compose_movetoutxo,
         ### /assets ###
         "/v2/assets": queries.get_valid_assets,
         "/v2/assets/<asset>": queries.get_asset,
