+++ counterparty-core/tools/compareledger.py
@@ -149,7 +149,7 @@ def get_last_block(database_file_1, database_file_2):
 database_file_1 = sys.argv[1]
 database_file_2 = sys.argv[2]
 
-LAST_BLOCK = 700000
+LAST_BLOCK = 650000
 # compare_ledger(database_file_1, database_file_2)
 check_hashes(database_file_1, database_file_2, "ledger_hash")
 # get_checkpoints(database_file_1)
